Revert common outbound Graphql
